+++ admin/setup_endpoints.php
@@ -1,0 +1,195 @@
+<?php
+/* Copyright (C) 2012-2022 Laurent Destailleur  <eldy@users.sourceforge.net>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ * or see https://www.gnu.org/
+ */
+
+/**
+ *     \file       htdocs/sellyoursaas/admin/setup_endpoints.php
+ *     \brief      Page administration module SellYourSaas (tab endpoints)
+ */
+
+
+if (! defined('NOSCANPOSTFORINJECTION')) define('NOSCANPOSTFORINJECTION', '1');		// Do not check anti CSRF attack test
+
+
+// Load Dolibarr environment
+$res=0;
+// Try main.inc.php into web root known defined into CONTEXT_DOCUMENT_ROOT (not always defined)
+if (! $res && ! empty($_SERVER["CONTEXT_DOCUMENT_ROOT"])) $res=@include $_SERVER["CONTEXT_DOCUMENT_ROOT"]."/main.inc.php";
+// Try main.inc.php into web root detected using web root caluclated from SCRIPT_FILENAME
+$tmp=empty($_SERVER['SCRIPT_FILENAME'])?'':$_SERVER['SCRIPT_FILENAME'];$tmp2=realpath(__FILE__); $i=strlen($tmp)-1; $j=strlen($tmp2)-1;
+while ($i > 0 && $j > 0 && isset($tmp[$i]) && isset($tmp2[$j]) && $tmp[$i]==$tmp2[$j]) { $i--; $j--; }
+if (! $res && $i > 0 && file_exists(substr($tmp, 0, ($i+1))."/main.inc.php")) $res=@include substr($tmp, 0, ($i+1))."/main.inc.php";
+if (! $res && $i > 0 && file_exists(dirname(substr($tmp, 0, ($i+1)))."/main.inc.php")) $res=@include dirname(substr($tmp, 0, ($i+1)))."/main.inc.php";
+// Try main.inc.php using relative path
+if (! $res && file_exists("../../main.inc.php")) $res=@include "../../main.inc.php";
+if (! $res && file_exists("../../../main.inc.php")) $res=@include "../../../main.inc.php";
+if (! $res) die("Include of main fails");
+
+require_once DOL_DOCUMENT_ROOT."/core/lib/admin.lib.php";
+require_once DOL_DOCUMENT_ROOT."/core/lib/files.lib.php";
+require_once DOL_DOCUMENT_ROOT."/core/lib/images.lib.php";
+require_once DOL_DOCUMENT_ROOT."/core/lib/geturl.lib.php";
+require_once DOL_DOCUMENT_ROOT."/core/class/html.formother.class.php";
+require_once DOL_DOCUMENT_ROOT."/core/class/html.formticket.class.php";
+require_once DOL_DOCUMENT_ROOT."/categories/class/categorie.class.php";
+dol_include_once('/sellyoursaas/lib/sellyoursaas.lib.php');
+dol_include_once('sellyoursaas/class/deploymentserver.class.php');
+
+// Access control
+if (! $user->admin) accessforbidden();
+
+// Parameters
+$action = GETPOST('action', 'aZ09');
+$backtopage = GETPOST('backtopage', 'alpha');
+
+$langs->loadLangs(array("admin", "errors", "install", "sellyoursaas@sellyoursaas"));
+
+//exit;
+
+// Initialize technical object to manage hooks of page. Note that conf->hooks_modules contains array of hook context
+$hookmanager->initHooks(array('sellyoursaas-setup'));
+
+$tmpservices=array();
+$staticdeploymentserver = new Deploymentserver($db);
+if (empty(getDolGlobalString('SELLYOURSAAS_OBJECT_DEPLOYMENT_SERVER_MIGRATION'))) {
+	$tmpservicessub = explode(',', getDolGlobalString('SELLYOURSAAS_SUB_DOMAIN_NAMES'));
+} else {
+	$tmpservicessub = $staticdeploymentserver->fetchAllDomains();
+}
+foreach ($tmpservicessub as $key => $tmpservicesub) {
+	$tmpservicesub = preg_replace('/:.*$/', '', $tmpservicesub);
+	if ($key > 0) $tmpservices[$tmpservicesub]=getDomainFromURL($tmpservicesub, 1);
+	else $tmpservices['0']=getDomainFromURL($tmpservicesub, 1);
+}
+$arrayofsuffixfound = array();
+foreach ($tmpservices as $key => $tmpservice) {
+	$suffix = '';
+	if ($key != '0') $suffix='_'.strtoupper(str_replace('.', '_', $tmpservice));
+
+	if (in_array($suffix, $arrayofsuffixfound)) continue;
+	$arrayofsuffixfound[$tmpservice] = $suffix;
+}
+
+
+/*
+ * Actions
+ */
+
+if ($action == 'set') {
+	$error=0;
+
+	if (! $error) {
+		dolibarr_set_const($db, "SELLYOURSAAS_SECURITY_KEY", GETPOST("SELLYOURSAAS_SECURITY_KEY", 'none'), 'chaine', 0, '', $conf->entity);
+	}
+
+	if (! $error) {
+		setEventMessages($langs->trans("SetupSaved"), null, 'mesgs');
+	}
+}
+
+
+/*
+ * View
+ */
+
+$formother=new FormOther($db);
+$form=new Form($db);
+$formticket = new FormTicket($db);
+
+$help_url="";
+llxHeader("", $langs->trans("SellYouSaasSetup"), $help_url);
+
+$linkback='<a href="'.($backtopage?$backtopage:DOL_URL_ROOT.'/admin/modules.php?restore_lastsearch_values=1').'">'.$langs->trans("BackToModuleList").'</a>';
+print_fiche_titre($langs->trans('SellYouSaasSetup'), $linkback, 'setup');
+
+$error=0;
+
+$head = sellyoursaas_admin_prepare_head();
+print dol_get_fiche_head($head, "setup_endpoints", "SellYouSaasSetup", -1, "sellyoursaas@sellyoursaas");
+
+print '<form enctype="multipart/form-data" method="POST" action="'.$_SERVER["PHP_SELF"].'" name="form_index">';
+print '<input type="hidden" name="token" value="'.newToken().'">';
+print '<input type="hidden" name="action" value="set">';
+
+print '<div class="div-table-responsive-no-min">'; // You can use div-table-responsive-no-min if you dont need reserved height for your table
+print '<table class="noborder centpercent">';
+print '<tr class="liste_titre">';
+print '<td>'.$langs->trans("Parameters").'</td><td>'.$langs->trans("Value").'</td>';
+print '<td><div class="float">'.$langs->trans("Examples").'</div><div class="floatright"><input type="submit" class="button buttongen" value="'.$langs->trans("Save").'"></div></td>';
+print "</tr>\n";
+
+print '<tr class="oddeven"><td>'.$langs->trans("SecurityKeyForPublicPages").' <span class="opacitymedium">(To protect the URL for Spam reporting webhooks)</spam></td>';
+print '<td>';
+print '<input class="minwidth300" type="text" name="SELLYOURSAAS_SECURITY_KEY" value="'.getDolGlobalString('SELLYOURSAAS_SECURITY_KEY').'">';
+print '</td>';
+print '<td><span class="opacitymedium small">123456abcdef</span></td>';
+print '</tr>';
+
+print '</table>';
+print '</div>';
+
+print '</table>';
+print '</div>';
+
+print "</form>\n";
+
+
+print "<br>";
+
+
+// Define $urlwithroot
+$urlwithouturlroot=preg_replace('/'.preg_quote(DOL_URL_ROOT, '/').'$/i', '', trim($dolibarr_main_url_root));
+$urlwithroot=$urlwithouturlroot.DOL_URL_ROOT;		// This is to use external domain name found into config file
+//$urlwithroot=DOL_MAIN_URL_ROOT;						// This is to use same domain name than current. For Paypal payment, we can use internal URL like localhost.
+
+/*
+var_dump(DOL_URL_ROOT);
+var_dump(dol_buildpath('/sellyoursaas/public/spamreport.php', 1));
+var_dump(DOL_MAIN_URL_ROOT);
+*/
+
+$message = '';
+$url = dol_buildpath('/sellyoursaas/public/spamreport.php', 3).'?key='.urlencode(getDolGlobalString('SELLYOURSAAS_SECURITY_KEY', 'KEYNOTDEFINED')).'[&mode=test]';
+$message .= img_picto('', 'object_globe.png').' '.$langs->trans("EndPointFor", "SpamReport", '{s1}');
+$message = str_replace('{s1}', '', $message);
+print $message.'<br>';
+print '<div class="urllink">';
+print '<input type="text" id="spamurl" class="quatrevingtpercent" value="'.$url.'">';
+print '<a href="'.dol_buildpath('/sellyoursaas/public/spamreport.php', 3).'?key='.urlencode(getDolGlobalString('SELLYOURSAAS_SECURITY_KEY', 'KEYNOTDEFINED')).'&mode=test" target="_blank" rel="noopener">';
+print img_picto('', 'download', 'class="paddingleft hideonsmartphone"');
+print '</a>';
+print '</div>';
+print ajax_autoselect("spamurl");
+
+print '<br><br>';
+
+/*
+$message='';
+$url='<a href="'.dol_buildpath('/sellyoursaas/myaccount/public/test.php', 3).'?key='.($conf->global->SELLYOURSAAS_SECURITY_KEY?urlencode($conf->global->SELLYOURSAAS_SECURITY_KEY):'...').'" target="_blank">'.dol_buildpath('/sellyoursaas/public/test.php', 3).'?key='.($conf->global->SELLYOURSAAS_SECURITY_KEY?urlencode($conf->global->SELLYOURSAAS_SECURITY_KEY):'KEYNOTDEFINED').'</a>';
+$message.=img_picto('', 'object_globe.png').' '.$langs->trans("EndPointFor", "Test", '{s1}');
+$message = str_replace('{s1}', $url, $message);
+print $message;
+
+print "<br>";
+*/
+
+//dol_fiche_end();
+
+
+llxfooter();
+
+$db->close();