+++ class/sellyoursaasutils.class.php
@@ -64,9 +64,10 @@ class SellYourSaasUtils
 	 * CAN BE A CRON TASK
 	 *
 	 * @param	int		$restrictonthirdpartyid		0=All qualified draft invoices, >0 = Restrict on qualified draft invoice of thirdparty.
+	 * @param	int		$maxtoprocess				0=All, >0 = Nb max of invoices to process
 	 * @return	int									0 if OK, <>0 if KO (this function is used also by cron so only 0 is OK)
 	 */
-	public function doValidateDraftInvoices($restrictonthirdpartyid = 0)
+	public function doValidateDraftInvoices($restrictonthirdpartyid = 0, $maxtoprocess = 0)
 	{
 		global $conf, $langs, $user;
 
@@ -97,11 +98,14 @@ class SellYourSaasUtils
 		$sql.= " AND f.total_ttc > 0";
 		if ($restrictonthirdpartyid > 0) $sql.=" AND f.fk_soc = ".((int) $restrictonthirdpartyid);
 		$sql.= " ORDER BY f.datef, f.rowid";
+		if ($maxtoprocess > 0) {
+			$sql.= $this->db->plimit($maxtoprocess);
+		}
 
 		$resql = $this->db->query($sql);
 		if ($resql) {
 			$num_rows = $this->db->num_rows($resql);
-			$i= 0;
+			$i = 0;
 			while ($i < $num_rows) {
 				$obj = $this->db->fetch_object($resql);
 				if ($obj && $invoice->fetch($obj->rowid) > 0) {
@@ -109,15 +113,15 @@ class SellYourSaasUtils
 
 					$invoice->fetch_thirdparty();
 
-					if ($invoice->thirdparty->array_options['manualcollection']) {
-						dol_syslog("This thirdparty has manual collection on, so we don't validate invoice");
+					if (!empty($invoice->thirdparty->array_options['manualcollection'])) {
+						dol_syslog("This thirdparty has manual collection on, so we don't validate the invoice");
 					} else {
 						$tmparray = $invoice->thirdparty->getOutstandingBills('customer');
 						if ($tmparray['opened'] > 0) {
 							dol_syslog("This thirdparty has already open invoices, so we don't validate any other invoices");     // So only 1 invoice is validated per thirdparty and pass
 						}
 
-						// Search contract linked to invoice
+						// Search contracts linked to the invoice we try to validate
 						$invoice->fetchObjectLinked(null, '', null, '', 'OR', 1, 'sourcetype', 1);
 
 						if (is_array($invoice->linkedObjects['contrat']) && count($invoice->linkedObjects['contrat']) > 0) {
@@ -151,6 +155,7 @@ class SellYourSaasUtils
 									$outputlangs->loadLangs(array('main', 'bills', 'products', 'users', 'sellyoursaas@sellyoursaas'));
 
 									// Set notes with the $contract->array_options['options_commentonqty']
+									dol_syslog("Check if we must update the public note with the comment on qty", LOG_DEBUG);
 									if (!empty($contract->array_options['options_commentonqty'])) {
 										$publicnoteofcontract = str_replace('User Accounts', $outputlangs->trans("ListOfUsers"), $contract->array_options['options_commentonqty']);
 										$newpublicnote = dol_concatdesc($invoice->note_public, $publicnoteofcontract);
@@ -158,7 +163,7 @@ class SellYourSaasUtils
 									}
 
 									// Check amount
-									if (!empty($conf->global->SELLYOURSAAS_MAX_MONTHLY_AMOUNT_OF_INVOICE)) {
+									if (getDolGlobalInt('SELLYOURSAAS_MAX_MONTHLY_AMOUNT_OF_INVOICE')) {
 										$amountofinvoice = $invoice->total_ht;
 										$monthfactor = 1;
 										if ($invoice->fk_fac_rec_source > 0) {
@@ -174,9 +179,9 @@ class SellYourSaasUtils
 										}
 										dol_syslog("doValidateDraftInvoices The invoice to validate has amount = ".$amountofinvoice." and come from recurring invoice with frequency ".$tmpinvoicerec->frequency."/".$tmpinvoicerec->unit_frequency." so a month factor of ".$monthfactor);
 										// Check amount with monthfactor is lower than $conf->global->SELLYOURSAAS_MAX_MONTHLY_AMOUNT_OF_INVOICE
-										if ($amountofinvoice >= ($conf->global->SELLYOURSAAS_MAX_MONTHLY_AMOUNT_OF_INVOICE * $monthfactor)) {
+										if ($amountofinvoice >= (getDolGlobalInt('SELLYOURSAAS_MAX_MONTHLY_AMOUNT_OF_INVOICE') * $monthfactor)) {
 											$error++;
-											$this->error = 'The invoice '.$invoice->ref." can't be validated: Amount ".$amountofinvoice." > ".$conf->global->SELLYOURSAAS_MAX_MONTHLY_AMOUNT_OF_INVOICE." * ".$monthfactor;
+											$this->error = 'The invoice '.$invoice->ref." can't be validated: Amount ".$amountofinvoice." > ".getDolGlobalInt('SELLYOURSAAS_MAX_MONTHLY_AMOUNT_OF_INVOICE')." * ".$monthfactor;
 											$this->errors[] = $this->error;
 											break;
 										}
@@ -185,7 +190,7 @@ class SellYourSaasUtils
 									$result = $invoice->validate($user);
 
 									if ($result > 0) {
-										$draftinvoiceprocessed[$invoice->id]=$invoice->ref;
+										$draftinvoiceprocessed[$invoice->id] = $invoice->ref;
 
 										// Now we build the PDF invoice
 										$hidedetails = (GETPOST('hidedetails', 'int') ? GETPOST('hidedetails', 'int') : (! empty($conf->global->MAIN_GENERATE_DOCUMENTS_HIDE_DETAILS) ? 1 : 0));
@@ -193,7 +198,7 @@ class SellYourSaasUtils
 										$hideref = (GETPOST('hideref', 'int') ? GETPOST('hideref', 'int') : (! empty($conf->global->MAIN_GENERATE_DOCUMENTS_HIDE_REF) ? 1 : 0));
 
 										$model_pdf = ($invoice->model_pdf ? $invoice->model_pdf : $invoice->modelpdf);
-										$ret = $invoice->fetch($id); // Reload to get new records
+										$ret = $invoice->fetch($invoice->id); // Reload to get new records
 
 										dol_syslog("GETPOST('lang_id','aZ09')=".GETPOST('lang_id', 'aZ09')." invoice->thirdparty->default_lang=".(is_object($invoice->thirdparty)?$invoice->thirdparty->default_lang:'invoice->thirdparty not defined')." newlang=".$newlang." outputlangs->defaultlang=".$outputlangs->defaultlang);
 
@@ -783,7 +788,7 @@ class SellYourSaasUtils
 
 	/**
 	 * Action executed by scheduler
-	 * Loop on invoice for customer with default payment mode Stripe and take payment/send email. Unsuspend if it was suspended (done by trigger BILL_CANCEL or BILL_PAYED).
+	 * Loop on each sale invoice with default payment mode Stripe and take payment/send email. Unsuspend if it was suspended (done by trigger BILL_CANCEL or BILL_PAYED).
 	 * CAN BE A CRON TASK
 	 *
 	 * @param	int		$maxnbofinvoicetotry    		Max number of payment to do (0 = No max)
@@ -911,6 +916,7 @@ class SellYourSaasUtils
 	/**
 	 * doTakePaymentStripeForThirdparty
 	 * Take payment/send email for a given thirdparty ID. Unsuspend if it was suspended (done by trigger BILL_CANCEL or BILL_PAYED).
+	 * Note: Some code has been implemented to manage Stripe SEPA mode=ban, but method used by batch for payment of direct debit is doTakePaymentStripeSEPA() not calling doTakePaymentStripeForThirdparty().
 	 *
 	 * @param	int		             $service					'StripeTest' or 'StripeLive'
 	 * @param	int		             $servicestatus				Service 0 or 1
@@ -941,6 +947,18 @@ class SellYourSaasUtils
 		// Check parameters
 		if (empty($thirdparty_id)) {
 			$this->errors[]='Empty parameter thirdparty_id when calling doTakePaymentStripeForThirdparty';
+			return 1;
+		}
+
+		if ($mode == 'ban') {
+			// This mode is not supported here. If you need it, you can do this process:
+
+			// Create the request to make a SEPA direct debit in database (add a record into prelevement_demande)
+			// $did = $invoice->demande_prelevement($user, 0, 'direct-debit', 'facture');
+			// Create the direct debit order
+			//$result = $invoice->makeStripeSepaRequest($user, $did);
+
+			$this->errors[] = 'Mode ban is not supported by doTakePaymentStripeForThirdparty(). You can do it with invoice->demande_prelevement() then invoice->makeStripeSepaRequest()';
 			return 1;
 		}
 
@@ -1184,7 +1202,8 @@ class SellYourSaasUtils
 							}
 						}
 
-						if (!$error) {	// Payment was not canceled
+						if (!$error) {
+							// Payment try can be done (no reason to have it canceled)
 							if ($mode == "ban") {
 								$stripecard = $stripe->sepaStripe($customer, $companypaymentmode, $stripeacc, $servicestatus, 0);
 							} elseif ($mode == "card") {
@@ -1231,7 +1250,7 @@ class SellYourSaasUtils
 										$stripefailuremessage=$e->getMessage();
 									}
 								} else { // Using new SCA method
-									dol_syslog("* Create payment on card ".$stripecard->id.", amounttopay=".$amounttopay.", amountstripe=".$amountstripe.", FULLTAG=".$FULLTAG, LOG_DEBUG);
+									dol_syslog("* Create payment on payment mode ".$stripecard->id.", amounttopay=".$amounttopay.", amountstripe=".$amountstripe.", FULLTAG=".$FULLTAG, LOG_DEBUG);
 
 									// Create payment intent and charge payment (because of confirmnow = true)
 									$confirmnow = true;
@@ -1348,13 +1367,13 @@ class SellYourSaasUtils
 
 									$ispostactionok = 1;
 
-									// Creation of payment line
+									// Creation of payment line in database if payment mode is not Direct Debit.
 									include_once DOL_DOCUMENT_ROOT.'/compta/paiement/class/paiement.class.php';
 									$paiement = new Paiement($this->db);
 									$paiement->datepaye     = $now;
 									$paiement->date         = $now;
 									if ($currencyCodeType == $conf->currency) {
-										$paiement->amounts      = array($invoice->id => $amounttopay);   // Array with all payments dispatching with invoice id
+										$paiement->amounts = array($invoice->id => $amounttopay);   // Array with all payments dispatching with invoice id
 									} else {
 										$paiement->multicurrency_amounts = array($invoice->id => $amounttopay);   // Array with all payments dispatching
 
@@ -1437,16 +1456,19 @@ class SellYourSaasUtils
 
 								$object = $invoice;
 
-								// Send emails
+								// Set the label of email to use by default when success
 								$labeltouse = 'InvoicePaymentSuccess';
 								$sendemailtocustomer = 1;
 
+								// Overwrite if an error was found
 								if (empty($charge) || $charge->status == 'failed') {
 									$labeltouse = 'InvoicePaymentFailure';
-									if ($noemailtocustomeriferror) $sendemailtocustomer = 0;		// $noemailtocustomeriferror is set when error already reported on myaccount screen
+									if ($noemailtocustomeriferror) {
+										$sendemailtocustomer = 0;		// $noemailtocustomeriferror is set when error already reported on myaccount screen
+									}
 								}
 
-								// Track an event
+								// Add an action event into database
 								if (empty($charge) || $charge->status == 'failed') {
 									$actioncode='PAYMENT_STRIPE_KO';
 									$extraparams=$stripefailurecode;
@@ -1518,7 +1540,8 @@ class SellYourSaasUtils
 						$extraparams='';
 					}
 
-					// Send email + create action after
+
+					// Send email to customer + record action after
 					if ($sendemailtocustomer && $labeltouse) {
 						dol_syslog("* Send email with result of payment - ".$labeltouse);
 
@@ -1682,7 +1705,7 @@ class SellYourSaasUtils
 
 	/**
 	 * Action executed by scheduler
-	 * Loop on invoice for customer with default payment mode Stripe sepa and send sepa direct debit ot stripe / sendmail.
+	 * Loop on sale invoice with default payment mode Stripe sepa and send sepa direct debit ot stripe / sendmail.
 	 * CAN BE A CRON TASK
 	 *
 	 * @param	int		$maxnbofinvoicetotry    		Max number of payment to do (0 = No max)
@@ -1785,14 +1808,14 @@ class SellYourSaasUtils
 						$sql .= " FROM ".MAIN_DB_PREFIX."prelevement_demande";
 						$sql .= " WHERE fk_facture = ".((int) $obj->rowid);
 						$sql .= " AND type = 'ban'"; // To exclude record saved by other online payments like credit card payments
-						$sql .= " AND traite = 0";
+						$sql .= " AND traite = 0";	// To not process payment request that were already converted into a direct debit or credit transfer order (Note: fk_prelevement_bons is also empty when traite = 0)
 						$rsql = $this->db->query($sql);
 						if ($rsql) {
 							$n = $this->db->num_rows($rsql);
 							if ($n != 1) {
 								$error++;
-								dol_syslog('Failed to create Stripe sepa request for invoice id = '.$obj->rowid.'. Too many direct debit order', LOG_ERR);
-								$this->errors[] = 'Failed to create Stripe sepa request for invoice id = '.$obj->rowid.'. Too many direct debit order found. We should have only 1.';
+								dol_syslog('Failed to create Stripe sepa request for invoice id = '.$obj->rowid.'. Not enough or too many request to pay with direct debit order. We should have only 1.', LOG_ERR);
+								$this->errors[] = 'Failed to create Stripe sepa request for invoice id = '.$obj->rowid.'. Not enough or too many request to pay with direct debit order. We should have only 1.';
 							} else {
 								$objd = $this->db->fetch_object($rsql);
 								$result = $invoice->makeStripeSepaRequest($user, $objd->rowid);
@@ -2174,7 +2197,7 @@ class SellYourSaasUtils
 					// Test if there is pending invoice
 					$object->fetchObjectLinked(null, '', null, '', 'OR', 1, 'sourcetype', 1);
 
-					if (is_array($object->linkedObjects['facture']) && count($object->linkedObjects['facture']) > 0) {
+					if (!empty($object->linkedObjects['facture']) && is_array($object->linkedObjects['facture']) && count($object->linkedObjects['facture']) > 0) {
 						usort($object->linkedObjects['facture'], "cmp");
 
 						//dol_sort_array($contract->linkedObjects['facture'], 'date');
@@ -3069,7 +3092,7 @@ class SellYourSaasUtils
 								// Now we force disable of recurring invoices
 								$object->fetchObjectLinked(null, '', null, '', 'OR', 1, 'sourcetype', 1);
 
-								if (is_array($object->linkedObjects['facturerec'])) {
+								if (!empty($object->linkedObjects['facturerec']) && is_array($object->linkedObjects['facturerec'])) {
 									foreach ($object->linkedObjects['facturerec'] as $idtemplateinvoice => $templateinvoice) {
 										// Disabled this template invoice
 										$res = $templateinvoice->setValueFrom('suspended', 1);
@@ -3080,7 +3103,7 @@ class SellYourSaasUtils
 								}
 
 								// Delete draft invoices linked to this thirdparty, after a successfull undeploy
-								if (is_array($object->linkedObjects['facture'])) {
+								if (!empty($object->linkedObjects['facture']) && is_array($object->linkedObjects['facture'])) {
 									foreach ($object->linkedObjects['facture'] as $idinvoice => $invoicetodelete) {
 										if ($invoicetodelete->statut == Facture::STATUS_DRAFT) {
 											if (preg_match('/\(.*\)/', $invoicetodelete->ref)) {
@@ -3557,9 +3580,14 @@ class SellYourSaasUtils
 					break;
 				}
 
-				$tmpold = explode('.', $object->oldcopy->ref_customer, 2);
-				$sldAndSubdomainold = $tmpold[0];
-				$domainnameold = (empty($tmpold[1]) ? '' :$tmpold[1]);
+				// Define old domain data for the 'rename' remoteaction
+				$sldAndSubdomainold = '';
+				$domainnameold = '';
+				if (!empty($object->oldcopy->ref_customer)) {
+					$tmpold = explode('.', $object->oldcopy->ref_customer, 2);
+					$sldAndSubdomainold = $tmpold[0];
+					$domainnameold = (empty($tmpold[1]) ? '' :$tmpold[1]);
+				}
 
 				$orgname = $contract->thirdparty->name;
 				$countryid = 0;
@@ -3627,7 +3655,6 @@ class SellYourSaasUtils
 				dol_syslog("password0salted=".$password0salted." passwordmd5salted=".$passwordmd5salted." passwordsha256salted=".$passwordsha256salted, LOG_DEBUG);
 
 				$conf->global->MAIN_SECURITY_SALT = '';
-				//dol_syslog("Using empty salt for __APPPASSWORDxxx__ variables : ".$conf->global->MAIN_SECURITY_SALT);
 				$password0 = dol_hash($password);	// deprecated. Depend on master setup.
 				$passwordmd5 = dol_hash($password, 'md5');
 				$passwordsha256 = dol_hash($password, 'sha256');
@@ -3704,7 +3731,8 @@ class SellYourSaasUtils
 				$tmppackage->targetsrcfile2 = make_substitutions($tmppackage->targetsrcfile2, $substitarray);
 				$tmppackage->targetsrcfile3 = make_substitutions($tmppackage->targetsrcfile3, $substitarray);
 
-				$automigrationtmpdir = $dirfortmpfiles."/automigration_".$object->socid.".tmp";
+
+				$automigrationtmpdir = $dirfortmpfiles."/automigration".($object->socid > 0 ? "_".$object->socid : "").".tmp";
 				$automigrationdocumentarchivename = (empty($object->array_options["automigrationdocumentarchivename"]) ? '' : $object->array_options["automigrationdocumentarchivename"]);
 				$dirforexampleforsources = (empty($object->array_options["dirforexampleforsources"]) ? '' : $object->array_options["dirforexampleforsources"]);
 				$laststableupgradeversion = (empty($object->array_options["laststableupgradeversion"]) ? '' : $object->array_options["laststableupgradeversion"]);
@@ -3931,7 +3959,7 @@ class SellYourSaasUtils
 					dol_syslog("passwordmd5salted=".$passwordmd5salted);
 
 					$conf->global->MAIN_SECURITY_SALT = '';
-					dol_syslog("Using empty salt for __APPPASSWORDxxx__ variables : ".$conf->global->MAIN_SECURITY_SALT);
+					dol_syslog("Using empty salt for __APPPASSWORDxxx__ variables");
 					$password0 = dol_hash($password);
 					$passwordmd5 = dol_hash($password, 'md5');
 					$passwordsha256 = dol_hash($password, 'sha256');
@@ -4278,7 +4306,7 @@ class SellYourSaasUtils
 		// So here, we can update this linked rec invoices
 		/*
 		foreach($arrayofrefreshedcontract as $refreshedcontract) {
-		 	 $refreshedcontract->array_options['options_commentonqty'] = join(', ', $arrayofcomment);
+			  $refreshedcontract->array_options['options_commentonqty'] = join(', ', $arrayofcomment);
 		}
 		// Then we must update this linked rec invoices
 		foreach($arrayofrefreshedrecinvoice as $refreshedrecinvoice) {