+++ myaccount/index.php
@@ -1291,6 +1291,8 @@ if ($action == 'updateurl') {
 		include_once DOL_DOCUMENT_ROOT.'/societe/class/companybankaccount.class.php';
 		include_once DOL_DOCUMENT_ROOT.'/core/lib/bank.lib.php';
 
+		dol_syslog("Record the SEPA payment mode from myaccount");
+
 		$companybankaccount = new CompanyBankAccount($db);
 		$companybankaccount->label = GETPOST('bankname', 'alphanohtml');
 		$companybankaccount->bank = GETPOST('bankname', 'alphanohtml');
@@ -1481,6 +1483,8 @@ if ($action == 'updateurl') {
 		$iban = '';
 		$bic = '';
 	} else {
+		dol_syslog("Record the credit card");
+
 		// Case of Credit or Debit card
 		$setupintentid = GETPOST('setupintentid', 'alpha');